+++ Lidgren.Network/NetRandom.Implementations.cs
@@ -237,16 +237,13 @@ namespace Lidgren.Network
 		/// </summary>
 		public static new readonly CryptoRandom Instance = new CryptoRandom();
 
-		private RandomNumberGenerator m_rnd = new RNGCryptoServiceProvider();
-
 		/// <summary>
 		/// Seed in CryptoRandom does not create deterministic sequences
 		/// </summary>
 		[CLSCompliant(false)]
 		public override void Initialize(uint seed)
 		{
-			byte[] tmp = new byte[seed % 16];
-			m_rnd.GetBytes(tmp); // just prime it
+			// Nada
 		}
 
 		/// <summary>
@@ -256,7 +253,7 @@ namespace Lidgren.Network
 		public override uint NextUInt32()
 		{
 			var bytes = new byte[4];
-			m_rnd.GetBytes(bytes);
+			RandomNumberGenerator.Fill(bytes);
 			return (uint)bytes[0] | (((uint)bytes[1]) << 8) | (((uint)bytes[2]) << 16) | (((uint)bytes[3]) << 24);
 		}
 
@@ -265,7 +262,7 @@ namespace Lidgren.Network
 		/// </summary>
 		public override void NextBytes(byte[] buffer)
 		{
-			m_rnd.GetBytes(buffer);
+			RandomNumberGenerator.Fill(buffer);
 		}
 
 		/// <summary>
@@ -273,9 +270,7 @@ namespace Lidgren.Network
 		/// </summary>
 		public override void NextBytes(byte[] buffer, int offset, int length)
 		{
-			var bytes = new byte[length];
-			m_rnd.GetBytes(bytes);
-			Array.Copy(bytes, 0, buffer, offset, length);
+			RandomNumberGenerator.Fill(buffer.AsSpan(offset, length));
 		}
 	}
 }